# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/layers/create": {
            "post": {
                "description": "创建一个新的层",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "层"
                ],
                "summary": "创建新层",
                "parameters": [
                    {
                        "description": "层信息",
                        "name": "layer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_layer.CreateLayerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_layer.CreateLayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/layers/delete-layer": {
            "delete": {
                "description": "根据层ID删除层",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "层"
                ],
                "summary": "删除层",
                "parameters": [
                    {
                        "type": "string",
                        "description": "层ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_layer.DeleteLayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/layers/get-all": {
            "get": {
                "description": "获取所有层的列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "层"
                ],
                "summary": "列出所有层",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "层类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "父层ID",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_layer.ListLayersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/layers/getinfo": {
            "get": {
                "description": "根据层ID获取层详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "层"
                ],
                "summary": "获取层信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "层ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_layer.GetLayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/layers/updateinfo": {
            "put": {
                "description": "更新层详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "层"
                ],
                "summary": "更新层信息",
                "parameters": [
                    {
                        "description": "更新层请求",
                        "name": "layer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_layer.UpdateLayerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_layer.UpdateLayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb_layer.CreateLayerRequest": {
            "type": "object",
            "properties": {
                "layer": {
                    "$ref": "#/definitions/pb_layer.Layer"
                }
            }
        },
        "pb_layer.CreateLayerResponse": {
            "type": "object",
            "properties": {
                "layer": {
                    "$ref": "#/definitions/pb_layer.Layer"
                }
            }
        },
        "pb_layer.DeleteLayerResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_layer.GetLayerResponse": {
            "type": "object",
            "properties": {
                "layer": {
                    "$ref": "#/definitions/pb_layer.Layer"
                }
            }
        },
        "pb_layer.Layer": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间 (Unix时间戳)",
                    "type": "integer"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "description": "层域ID",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "string"
                },
                "ratio": {
                    "description": "占有父级比例",
                    "type": "number"
                },
                "type": {
                    "description": "类型（层或域）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pb_layer.LayerType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "更新时间 (Unix时间戳)",
                    "type": "integer"
                }
            }
        },
        "pb_layer.LayerType": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-comments": {
                "LayerType_DOMAIN": "域",
                "LayerType_LAYER": "层"
            },
            "x-enum-varnames": [
                "LayerType_LAYER",
                "LayerType_DOMAIN"
            ]
        },
        "pb_layer.ListLayersResponse": {
            "type": "object",
            "properties": {
                "layers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb_layer.Layer"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "pb_layer.UpdateLayerRequest": {
            "type": "object",
            "properties": {
                "layer": {
                    "$ref": "#/definitions/pb_layer.Layer"
                }
            }
        },
        "pb_layer.UpdateLayerResponse": {
            "type": "object",
            "properties": {
                "layer": {
                    "$ref": "#/definitions/pb_layer.Layer"
                }
            }
        },
...
    }
}